Move ReceptorVerifiyFunc to tlsconfig.go and write unit tests
