+++ pkg/netceptor/tlsconfig.go
@@ -1,6 +1,9 @@
 package netceptor
 
 import (
+	"bytes"
+	"crypto/sha256"
+	"crypto/sha512"
 	"crypto/tls"
 	"crypto/x509"
 	"encoding/hex"
@@ -8,7 +11,9 @@ import (
 	"fmt"
 	"os"
 	"strings"
+	"time"
 
+	"github.com/ansible/receptor/pkg/logger"
 	"github.com/ansible/receptor/pkg/utils"
 	"github.com/ghjm/cmdline"
 	"github.com/spf13/viper"
@@ -234,6 +239,183 @@ func (cfg TLSClientConfig) Prepare() error {
 	}
 
 	return MainInstance.SetClientTLSConfig(cfg.Name, tlscfg, pinnedFingerprints)
+}
+
+// **************************************************************************
+// Certificate Verification
+// **************************************************************************
+
+// ReceptorCertNameError represents an error when a certificate doesn't match expected Receptor node IDs.
+type ReceptorCertNameError struct {
+	ValidNodes   []string
+	ExpectedNode string
+}
+
+func (rce ReceptorCertNameError) Error() string {
+	if len(rce.ValidNodes) == 0 {
+		return fmt.Sprintf("x509: certificate is not valid for any Receptor node IDs, but wanted to match %s",
+			rce.ExpectedNode)
+	}
+	var plural string
+	if len(rce.ValidNodes) > 1 {
+		plural = "s"
+	}
+
+	return fmt.Sprintf("x509: certificate is valid for Receptor node ID%s %s, not %s",
+		plural, strings.Join(rce.ValidNodes, ", "), rce.ExpectedNode)
+}
+
+// VerifyType indicates whether we are verifying a server or client.
+type VerifyType int
+
+const (
+	// VerifyServer indicates we are the client, verifying a server.
+	VerifyServer VerifyType = 1
+	// VerifyClient indicates we are the server, verifying a client.
+	VerifyClient = 2
+)
+
+// ExpectedHostnameType indicates whether we are connecting to a DNS hostname or a Receptor Node ID.
+type ExpectedHostnameType int
+
+const (
+	// ExpectedHostnameTypeDNS indicates we are expecting a DNS style hostname.
+	ExpectedHostnameTypeDNS ExpectedHostnameType = 1
+	// ExpectedHostnameTypeReceptor indicates we are expecting a Receptor node ID.
+	ExpectedHostnameTypeReceptor = 2
+)
+
+// ReceptorVerifyFunc generates a function that verifies a Receptor node ID.
+func ReceptorVerifyFunc(tlscfg *tls.Config, pinnedFingerprints [][]byte, expectedHostname string,
+	expectedHostnameType ExpectedHostnameType, verifyType VerifyType, logger *logger.ReceptorLogger,
+) func(rawCerts [][]byte, verifiedChains [][]*x509.Certificate) error {
+	return func(rawCerts [][]byte, verifiedChains [][]*x509.Certificate) error {
+		if len(rawCerts) == 0 {
+			logger.Error("RVF failed: peer certificate missing")
+
+			return fmt.Errorf("RVF failed: peer certificate missing")
+		}
+		certs := make([]*x509.Certificate, len(rawCerts))
+		for i, asn1Data := range rawCerts {
+			cert, err := x509.ParseCertificate(asn1Data)
+			if err != nil {
+				logger.Error("RVF failed to parse: %s", err)
+
+				return fmt.Errorf("failed to parse certificate from server: " + err.Error()) //nolint:govet
+			}
+			certs[i] = cert
+		}
+		var opts x509.VerifyOptions
+		switch verifyType {
+		case VerifyServer:
+			opts = x509.VerifyOptions{
+				Intermediates: x509.NewCertPool(),
+				Roots:         tlscfg.RootCAs,
+				CurrentTime:   time.Now(),
+				KeyUsages:     []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth},
+			}
+			if expectedHostnameType == ExpectedHostnameTypeDNS && expectedHostname != "" {
+				opts.DNSName = expectedHostname
+			}
+		case VerifyClient:
+			opts = x509.VerifyOptions{
+				Intermediates: x509.NewCertPool(),
+				Roots:         tlscfg.ClientCAs,
+				CurrentTime:   time.Now(),
+				KeyUsages:     []x509.ExtKeyUsage{x509.ExtKeyUsageClientAuth},
+			}
+			if expectedHostnameType == ExpectedHostnameTypeDNS && expectedHostname != "" {
+				opts.DNSName = expectedHostname
+			}
+		default:
+			logger.Error("RVF failed: invalid verification type: must be client or server")
+
+			return fmt.Errorf("RVF failed: invalid verification type: must be client or server")
+		}
+
+		if len(pinnedFingerprints) > 0 {
+			var sha224sum []byte
+			var sha256sum []byte
+			var sha384sum []byte
+			var sha512sum []byte
+			fingerprintOK := false
+			for _, fing := range pinnedFingerprints {
+				fingLenFound := false
+				for _, s := range []struct {
+					len     int
+					sum     *[]byte
+					sumFunc func(data []byte) []byte
+				}{
+					{28, &sha224sum, func(data []byte) []byte {
+						sum := sha256.Sum224(data)
+
+						return sum[:]
+					}},
+					{32, &sha256sum, func(data []byte) []byte {
+						sum := sha256.Sum256(data)
+
+						return sum[:]
+					}},
+					{48, &sha384sum, func(data []byte) []byte {
+						sum := sha512.Sum384(data)
+
+						return sum[:]
+					}},
+					{64, &sha512sum, func(data []byte) []byte {
+						sum := sha512.Sum512(data)
+
+						return sum[:]
+					}},
+				} {
+					if len(fing) == s.len {
+						fingLenFound = true
+						if *s.sum == nil {
+							*s.sum = s.sumFunc(certs[0].Raw)
+						}
+						if bytes.Equal(fing, *s.sum) {
+							fingerprintOK = true
+
+							break
+						}
+					}
+				}
+				if !fingLenFound {
+					logger.Error("RVF failed: pinned certificate must be sha224, sha256, sha384 or sha512")
+
+					return fmt.Errorf("RVF failed: pinned certificate must be sha224, sha256, sha384 or sha512")
+				}
+			}
+			if !fingerprintOK {
+				logger.Error("RVF failed: presented certificate does not match any pinned fingerprint")
+
+				return fmt.Errorf("RVF failed: presented certificate does not match any pinned fingerprint")
+			}
+		}
+
+		for _, cert := range certs[1:] {
+			opts.Intermediates.AddCert(cert)
+		}
+		_, err := certs[0].Verify(opts)
+		if err != nil {
+			logger.Error("RVF failed verify: %s\nRootCAs: %v\nServerName: %s", err, tlscfg.RootCAs, tlscfg.ServerName)
+
+			return err
+		}
+
+		if expectedHostnameType == ExpectedHostnameTypeReceptor {
+			found, receptorNames, err := utils.ParseReceptorNamesFromCert(certs[0], expectedHostname, logger)
+			if err != nil {
+				return err
+			}
+			if !found {
+				logger.Error("RVF ReceptorNameError: expected %s but found %s", expectedHostname, strings.Join(receptorNames, ", "))
+
+				return ReceptorCertNameError{ValidNodes: receptorNames, ExpectedNode: expectedHostname}
+			}
+		}
+
+		return nil
+	}
 }
 
 func init() {
